+++ src/commands/generate_access_token.rs
@@ -1,8 +1,9 @@
 use crate::opt::GenerateAccessTokenInput;
+use crate::result::Result;
 use imgur_openapi::apis::auth_api;
 use imgur_openapi::apis::configuration::Configuration;
 
-pub async fn generate_access_token(input: GenerateAccessTokenInput) {
+pub async fn generate_access_token(input: GenerateAccessTokenInput) -> Result<()> {
     let configuration = Configuration::new();
     let model = auth_api::generate_access_token(
         &configuration,
@@ -11,8 +12,8 @@ pub async fn generate_access_token(input: GenerateAccessTokenInput) {
         "refresh_token",
         &input.refresh_token,
     )
-    .await
-    .unwrap();
-    let json = serde_json::to_string(&model).unwrap();
+    .await?;
+    let json = serde_json::to_string(&model)?;
     println!("{}", json);
+    Ok(())
 }