+++ src/error.rs
@@ -1,17 +1,42 @@
+use imgur_openapi::apis;
+use std::io;
+
 #[derive(Debug)]
-pub enum Error<T> {
-    JsonSerializeError(serde_json::Error),
-    ApiError(imgur_openapi::apis::Error<T>),
+pub enum Error {
+    ApiOthers,
+    ApiResponse(Response),
+    Io(io::Error),
+    JsonSerialize(serde_json::Error),
 }
 
-impl<T> From<serde_json::Error> for Error<T> {
-    fn from(source: serde_json::Error) -> Self {
-        Error::JsonSerializeError(source)
+#[derive(Debug)]
+pub struct Response {
+    content: String,
+    status: u16,
+}
+
+impl From<io::Error> for Error {
+    fn from(source: io::Error) -> Self {
+        Error::Io(source)
     }
 }
 
-impl<T> From<imgur_openapi::apis::Error<T>> for Error<T> {
-    fn from(source: imgur_openapi::apis::Error<T>) -> Self {
-        Error::ApiError(source)
+impl From<serde_json::Error> for Error {
+    fn from(source: serde_json::Error) -> Self {
+        Error::JsonSerialize(source)
+    }
+}
+
+impl<T> From<apis::Error<T>> for Error {
+    fn from(source: apis::Error<T>) -> Self {
+        match source {
+            apis::Error::ResponseError(apis::ResponseContent {
+                status, content, ..
+            }) => Error::ApiResponse(Response {
+                status: status.as_u16(),
+                content,
+            }),
+            _ => Error::ApiOthers,
+        }
     }
 }