+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod commands;
 pub mod error;
 pub mod opt;
+pub mod result;